Refactor to add an interface for view_state of #mrheader
The goal of this refactoring is to make it easier to understand and maintain
view_state data structure of #mrheader.

Unlike record #mrview in #mrst, the data structure that represents
view_states in #mrheader is a tuple (of tuples) and its defenition
and application spread across all couch_mrview_util.erl module.
This makes it hard to follow and understand how it used
or troubleshoot mrview in case of the problems.

This change introduces an interface for generation and accessing
view_state and also separates generation of reduction clousers
from mrview's `reduce`, `reduce_to_count` and `fold_reduce` functions,
as they rely on view_state's structure.