Upgrade/downgrade mrargs record
This is a fixup commit for COUCHDB-3063.
The issue appear when we run a mixed cluster when some nodes has old
version of record and others upgraded already.
We upgrade mrargs record to include version field to be able to
distinguish different versions of the record. Which would help us
going forward.

COUCHDB-3063